requirements.txt
wand
pyyaml
discord.py
thefuzz
